requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
bcrypt==4.0.1
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
ecdsa==0.18.0
elasticsearch==7.10.1
exceptiongroup==1.1.3
fastapi==0.103.2
frozenlist==1.4.0
h11==0.14.0
idna==3.4
multidict==6.0.4
passlib==1.7.4
pyasn1==0.5.0
pydantic==2.4.2
pydantic_core==2.10.1
python-dotenv==1.0.0
python-jose==3.3.0
requests==2.31.0
requests-aws4auth==1.2.3
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.8.0
urllib3==1.26.17
uvicorn==0.23.2
yarl==1.9.2
